Change: format/delete-invisible-text now treats invisibility by overlays
Previously, neos-live-editor/format/delete-invisible-text deleted
invisible texts marked by text properties but leave texts marked by
overlays as is.

It was because I'm using 'original-buffer's overlays as reference
instead of copying it to the temporary buffer (thus, I needed to
pass it as argument right now).

From this commit, I modified it to also delete invisible texts marked
by overlays too. This is what it is expected from its name.